# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ISGIDocs API",
        "description": "ISGIDocs est une API conçue pour gérer les ressources pédagogiques et les formations de l'Institut Supérieur de Génie Informatique. Cette API permet aux utilisateurs de gérer les modules, les documents pédagogiques, et l'authentification des utilisateurs pour une expérience d'auto-formation fluide et organisée.",
        "contact": {
            "email": "salahaznidi09@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/modules": {
            "get": {
                "tags": [
                    "Module"
                ],
                "summary": "Liste tous les modules",
                "operationId": "348e9bb569091f8a261998a3b6685d5a",
                "responses": {
                    "200": {
                        "description": "Liste des modules récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Module"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Module"
                ],
                "summary": "Ajoute un nouveau module",
                "operationId": "29f95790bb17b4f411e97e6fde9ccbe9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "MATH101"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Mathématiques 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Introduction aux mathématiques"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Module ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Module"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données non valides"
                    }
                }
            }
        },
        "/api/modules/{id}": {
            "get": {
                "tags": [
                    "Module"
                ],
                "summary": "Récupère un module par ID",
                "operationId": "53696906f8a449aabec698e9b9d23b75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du module",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du module récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Module"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Module non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Module"
                ],
                "summary": "Modifie un module existant",
                "operationId": "0446fed3c1bef7b517980c1eb97eae47",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du module à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "MATH101"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Mathématiques avancées"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cours avancé de mathématiques"
                                    },
                                    "annee": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Module modifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Module"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Module non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Module"
                ],
                "summary": "Supprime un module",
                "operationId": "613f8e5cfa8d75c4da11dd09a0034ffa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du module à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Module supprimé avec succès"
                    },
                    "404": {
                        "description": "Module non trouvé"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Module": {
                "title": "Module",
                "description": "Représente un module de formation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "MATH101"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Mathématiques 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Introduction aux mathématiques"
                    },
                    "annee": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Entrer un token d'accès après avoir été authentifié.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}